Automatically handle native character sets on input to the mail utility.
If the mime header is set, then mail will provide the missing 'charset'
parameter for each Content-Type header that begins with 'text/'.
Its value will be determined by examining the 'charset' mail variable.
If it is set to 'auto' (the default), the character set will be extracted
from the value of the LC_ALL environment variable.  If it is unset, it
will be deduced from the LANG environment variable.

Thus, provided that LC_ALL is set correctly, the following setting in
.mailrc is recommended to ensure that mails in native character
sets will be processed correctly:

  set charset=auto mime

In most cases, it can be simplified to just 'set mime'.

* NEWS: Update.
* doc/texinfo/programs.texi: Update the description of the charset
variable.
* mail/mail.h (util_get_charset): New proto.
* mail/send.c (attach_set_content_type): New function.
(attlist_add, add_body): Use attach_set_content_type to
set the content_type field.
* mail/util.c (util_get_charset): New function.
(util_rfc2047_decode): Use util_get_charset.