Fix the operation of CRLF and CRLFDOT filters.
When encoding, these filters used to retain any existing CRLF
sequences untouched on output. As a result, it was not possible to
predict the number of characters output having the number of
characters and number of lines on input. Due to this imap4d fetch
would produce invalid literal output on FETCH command if the
requested stream (usually body), contained a mixture of LF and
CRLF terminated lines. Moreover, encode + decode operations would
not be idempotent in this case.

To fix this, the default behavior of CRLF and CRLFDOT filter in
encode mode is changed to translate each LF to CRLF, no matter what
the prior character was. To invoke the former "normalizing" behavior,
the "-n" option argument is provided for mu_filter_create_args.

* imap4d/fetch.c (fetch_io): Use raw stream when possible.
Use a temporary CRLF translator to ensure proper positioning if
start offset is specified. This is suboptimal (to say the very
least) and will be fixed as soon as possible.
* imap4d/io.c (io_enable_crlf): New funcition (hopefully temporary).
* libmailutils/filter/crlfdot.c: When encoding, retain input CRLF
only if the -n option was used when creating the filter. Otherwise,
translate it to CRCRLF.
* libmailutils/filter/crlfflt.c: Likewise.
* libmailutils/stream/fltstream.c (filter_read): Make sure
mu_filter_lastbuf is always emitted prior to closing the filter.
(filter_seek): Clear flag_eof.
* libmailutils/tests/Makefile.am (noinst_PROGRAMS): Add dump and tocrlf.
* libmailutils/tests/crlf.at: New test case.
* libmailutils/tests/crlfdot.at: New test case.
* libmailutils/tests/dump.c: New auxiliary program.
* libmailutils/tests/tocrlf.c: Likewise.
* libmailutils/tests/testsuite.at (MU_FILTER_TEST_NIBBLE)
(MU_FILTER_TEST): New defines.