mail: new notation @, for the mailbox given with the -f option
* doc/texinfo/programs/mail.texi: Document the @ notation.
* NEWS: Likewise.
* mail/copy.c (mail_copy0): Rewrite. Use mailbox URLs instead of
filenames.
* mail/file.c (mail_expand_name): Return integer code. On success,
convert the expanded name to URL and return it in the memory location
supplied by the second argument. Use new shorthand notation @, for
the name of the file given with the -f option.
All uses changed.
* mail/mail.c (secondary_url): New global.
(main): Save the URL of the file given with the -f option in
secondary_url. Use it, instead of the file variable, to decide
what kind of summary message to display for empty mailbox (see
commit 99b6b6f02c3cf94ebe1daf7233a702c743d6a2b5).
* mail/mail.h (secondary_url): New global.
(mail_expand_name): Change prototype.
* mail/mailline.c: Minor changes.

* libmailutils/wicket/noauth.c (noauth_ticket_get_cred): Return
MU_ERR_AUTH_NO_CRED.