Split maidag into three single-purpose tools
Maidag operated in three modes: mda, url and lmtp.  In mda mode
it is normally started by sendmail with non-root privileges.
In order to be able to write to other users' mailboxes, the binary
must be setuid, which creates certain security threats.  As these appear
to be impossible to avoid in a single multi-purpose program, it has
been decided to replace maidag with three single-purpose tools: mda,
putmail, and lmtpd.  The mda tool is a local mail delivery agent.  It is
the only one for which the setuid bit is required, so special measures
have been taken to secure it.  In particular, to avoid unverified user
inputs it disallows to override any configuration settings from the
command line.

The two other tools provide the functionality of the url and lmtp mode.

* Makefile.am (MAIDAG_DIR): Replace with MDA_DIR
* configure.ac (MU_COND_MAILBOX_QUOTA): New conditional.
Use names relative to ${top_builddir} instead of ".".
* maidag: Remove.
* mda/.gitignore: New file.
* mda/Makefile.am: New file.
* mda/lib/Makefile.am: New file.
* mda/lib/deliver.c: New file.
* mda/lib/forward.c: New file.
* mda/lib/libmda.h: New file.
* mda/lib/mailquota.c: New file.
* mda/lib/script.c: New file.
* mda/lib/util.c: New file.
* mda/lmtpd.c: New file.
* mda/mda.c: New file.
* mda/putmail.c: New file.
* mda/tests/.gitignore: New file.
* mda/tests/Makefile.am: New file.
* mda/tests/atlocal.in: New file.
* mda/tests/input.msg: New file.
* mda/tests/lmtpd.at: New file.
* mda/tests/putmail.at: New file.
* mda/tests/testsuite.at: New file.

* doc/texinfo/programs/maidag.texi: Remove.
* doc/texinfo/programs/mda.texi: New file.
* doc/texinfo/programs/lmtpd.texi: New file.