Amend b8664765e1
The body_end member holds the offset of the first character
that follows the message body.  In a well-formed mailbox, it is
the newline character of a blank line that delimits the message.
However, there are mailboxes that lack this blank line.  In such
cases body_end points to the ^F of the next envelope line.

When saving the mailbox, delimiting blank lines are always emitted.

* libproto/mbox/mbox.c (mbox_expunge_unlocked): Revert size calculation.
Always emit a blank line after the message.