Fix memory leak in property code
The property API contains a serious flaw that leads to memory leaks.
The property initialization function (set by mu_property_set_init or
similiar functions) is supposed to initialize the supported methods
in mu_property_t, in particular _prop_done, which is responsible for
releasing the resources allocated for the initdata object upon
invocation of mu_property_destroy.

However, the initialization function gets called only upon a read/write
manipulation on the property. This means that unless at least one
such operation is performed the _prop_done method will not be initialized
and the resources (if any) allocated for the initial data object will
never be released.  This is true, in particular for the MH properties
and assoc properties with I/O stream.

Until the API is revamped, the following solution is implemented.
The function mu_property_create_init calls the mu_property_init
function if called with a non-NULL third argument (initdata).
This ensures that it will be properly freed upon destroying the
object.

The use of low-level functions mu_property_set_init and
mu_property_set_init_data requires the caller to call mu_property_init
explicitly, if needed.

* include/mailutils/property.h (mu_property_init): New proto.
* libmailutils/property/create.c (_mu_property_init): Rename to
mu_property_init.
(mu_property_create_init): Call mu_property_init if initdata is not
NULL.
* libmailutils/property/propinv.c: Update call to mu_property_init.