Relax sytnax requirements for MIME structured headers.
* libmailutils/base/ctparse.c (content_type_parse): Rewrite
the value splitting on type and subtype.  Allow for optional
whitespace at both sides of "/".  Both type and subtype can contain
arbitrary characters (except for "/").  This is not right according
to RFC 2045, but reportedly such cases exist in old mails.
* libmailutils/mailbox/bodystruct.c (bodystructure_fill): Fix
parsing of the Content-Type header.
Fix storing the pointer to an automatic variable in assoc object.
Unfold the Content-Disposition value.
* libmailutils/mime/mimehdr.c (_mime_header_parse): Return in pvalue
entire prefix part up to the first semicolon, with leading and
trailing whitespace removed.  Allow for both output parameters to
be NULL.
* libmailutils/tests/content-type.at: Add new test case.
* libmailutils/tests/conttype.c: ignore empty lines in input.
* libmailutils/tests/mimehdr.at: Change mimehdr16 and mimehdr17
tests: this syntax is accepted by the relaxed rules of the
modified parser.