Implement temporary streams.
A temporary stream functions as a memory stream until its size reaches
a preconfigured threshold value.  Once it is reached, the stream storage
is automatically converted to a temporary file, and all data written so far
are transferred to the new storage.  If the temporary file cannot be
created, the stream continues to operate in memory-based mode.

This stream type is intended to decrease the number of used file
descriptors.

* include/mailutils/sys/Makefile.am: Add temp_stream.h
* include/mailutils/sys/temp_stream.h: New file.
* include/mailutils/stream.h (mu_temp_file_threshold_size): New extern.
(mu_temp_stream_create): New proto.
* libmailutils/stream/Makefile.am: Add temp_stream.c.
* libmailutils/stream/temp_stream.c: New file.
* libmailutils/string/Makefile.am: Add strtosize.c

* libmailutils/tests/.gitignore: Update.
* libmailutils/tests/Makefile.am
* libmailutils/tests/strtoc.c: Add tests for mu_c_hsize.
* libmailutils/tests/temp_stream.c: New test program.
* libmailutils/tests/testsuite.at: Test the temp_stream functionality.
* testsuite/testsuite.inc (MU_CHECK): New macro.

* libmailutils/mailbox/body.c (body_get_transport): Use mu_temp_stream_create.