Reorganize testsuite.
The testsuite directory contains only data and programs that are
used by other tests in the mailutils hierarchy.  As such, it is
moved to the top of the SUBDIRS list, so that the utilities on which
another tests depend are tested first.

The tests that depend on both libmailutils and libraries from libproto
are moved to a separate directory libtests.  These tests are run after
tests in testsuite, libmailutils and libproto.

* Makefile.am (SUBDIR): Move the testsuite directory before libmailutils.
* configure.ac (libtests): New test directory.
* libtests/.gitignore: New file.
* libtests/Makefile.am: New file.
* libtests/atlocal.in: New file.
* testsuite/fldel.at: Move to libtests/
* testsuite/fldel.c: Likewise.
* testsuite/lstuid.c: Likewise.
* testsuite/lstuid00.at: Likewise.
* testsuite/lstuid01.at: Likewise.
* testsuite/lstuid02.at: Likewise.
* testsuite/mbdel.at: Likewise.
* testsuite/mbdel.c: Likewise.
* testsuite/mime.at: Likewise.
* testsuite/mime01.at: Likewise.
* testsuite/mimetest.c: Likewise.
* libtests/testsuite.at: New file.
* testsuite/Makefile.am: Remove all files that depend on mailutils
components.
* testsuite/testsuite.at: Likewise.
* testsuite/bs.c: Remove.
* testsuite/msgset.c: Remove.