Don't assume any special meaning of the X-Imapbase header in the library.
This starts a series of commits whose purpose is to clean up the
mess with the uid and uidvalidity attributes.

The way these attributes are stored depends on the mailbox type.  The
libmailutils library contained some historical left-overs that assumed
that the X-Imapbase header of the first message in the mailbox contained
the value of the uidvalidity and next uid.  These left-overs are removed.

* include/mailutils/message.h (mu_message_scan): Remove uidvalidity.
* libmailutils/mailbox/msgscan.c: Remove special handling for the
X-Imapbase header.
* libmailutils/base/amd.c (_amd_message_save)
(amd_scan_message): Remove special handling for X-Imapbase.