Full rewrite of maildir support
This rewrite aims to speed up maildir manipulations and to fix the
problem with storing the uids and uidvalidity attributes.

Uidvalidity value is stored in the .mu-prop file, as it was done
previously.  UID of each message is encoded in its file name
between the unique prefix and the info part.  Overall structure
of the message file name is:

  uniq [ ',' attrs ] ':2,' info

where uniq is the unique prefix, info is the attribute flags, and attrs
is a comma-delimited list of attribute=value pairs.  The 'u' attribute
holds the UID.  Unrecognized attributes are ignored.

To speed up the access and avoid unnecessary memory allocations, maildir
access is reimplemented using the *at() function calls (openat instead
of open, etc).

* libproto/maildir/folder.c: Rewrite.
* libproto/maildir/maildir.h: Rewrite.
* libproto/maildir/mbox.c: Rewrite.