mu_scan_datetime: Fix error recovery.  %Z handles both time-zone forms.
The value stored in the memory location pointed to by the endp argument
has changed.  On the MU_ERR_PARSE error, it is (as earlier) the position
in input where conversion stopped.  On the MU_ERR_FORMAT, however, it
is the character in fmt where the format error was detected.

* libmailutils/datetime/scantime.c (peek_state): Return MU_ERR_FORMAT
if mu_list_tail fails (the list is empty).  This reflects unbalanced
closing parenthesis or bracket.
(scan_recovery): fix algorithm.
(mu_scan_datetime): %Z handles both abbreviated time-zones and
time-zones as hour offset from GMT.
Improve error handling.
On MU_ERR_PARSE error, point *endp to the character in input where
conversion stopped.
On MU_ERR_FORMAT error, point it to the character in fmt, where the
format error was detected.
* libmailutils/tests/scantime.c: Improve error handling.
* libmailutils/tests/scantime.at: More testcases.