Optimize the observable interface
A bitmask of all requested event types is stored in the observable
structure.  When an event is reported, mu_observable_notify first
check if the event type bit is set in that bitmap.  If so, it starts
iterating over observers.  That should allow us to avoid the loop in
most cases.

* include/mailutils/observer.h (mu_observable_wants): New proto.
* include/mailutils/sys/observer.h (_mu_observable): New member: types.
* libmailutils/base/observer.c (mu_observable_attach): Compute bitmask
of all types requested by attached observers.
(mu_observable_detach): Likewise.
(mu_observable_notify): Check if the event type is marked in the bitmap,
skip iteration if it is not.
(mu_observable_wants): New function.