mbox: fix expunging in a read-only directory
This fixes debian bug 980042.
Re: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=980042#16

* libproto/mbox/mboxrd.c (mboxrd_flush_unlocked): If temporary file
cannot be created in the directory of the mailbox because of EACCES,
create it elsewhere and copy back to the mailbox when ready.
Also, make sure that the mode and ownership of the mailbox are preserved,
no matter what temporary file was used.
* libproto/mbox/tests/rospool.at: New test.
* libproto/mbox/tests/Makefile.am: Add new test.
* libproto/mbox/tests/testsuite.at: Add new test.